Encapsulated some image methods in a separate class
