+++ InventoryManager/app/src/main/java/com/example/inventorymanager/ui/editItem/EditItemFragment.java
@@ -39,6 +39,7 @@ import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.target.SimpleTarget;
 import com.bumptech.glide.request.transition.Transition;
 import com.example.inventorymanager.ImageSelectionFragment;
+import com.example.inventorymanager.ImageUtility;
 import com.example.inventorymanager.Item;
 import com.example.inventorymanager.ItemUtility;
 import com.example.inventorymanager.ItemViewModel;
@@ -75,7 +76,7 @@ public class EditItemFragment extends Fragment {
     private ArrayList<String> imagePaths = new ArrayList<>(); // local paths
     private ArrayList<String> tempList = new ArrayList<String>(); // A temporary list to store our generated urls
 
-
+    private ImageUtility imageUtility;
     private ImageView imageView0;
     private Button addImage0Button;
     private Button deleteImage0Button;
@@ -86,11 +87,8 @@ public class EditItemFragment extends Fragment {
     private ImageView imageView2;
     private Button addImage2Button;
     private Button deleteImage2Button;
-
-    private static final int REQUEST_CAMERA_PERMISSION = 1;
     private static final int REQUEST_CAMERA = 2;
-    private static final int REQUEST_GALLERY_PERMISSION = 3;
-    private static final int REQUEST_GALLERY = 4;
+    private static final int REQUEST_GALLERY = 3;
 
     /**
      * Provides the user interface of the fragment.
@@ -239,17 +237,10 @@ public class EditItemFragment extends Fragment {
 
         // IMAGE FUNCTIONALITY
         // when you click the respective Add Image button, choose if you're gonna add from gallery or take a pic with camera
-        addImage0Button.setOnClickListener( v -> {
-            showImageOptionsDialog();
-        });
-
-        addImage1Button.setOnClickListener( v -> {
-            showImageOptionsDialog();
-        });
-
-        addImage2Button.setOnClickListener( v -> {
-            showImageOptionsDialog();
-        });
+        imageUtility = new ImageUtility(this);
+        addImage0Button.setOnClickListener(v -> imageUtility.showImageOptionsDialog());
+        addImage1Button.setOnClickListener(v -> imageUtility.showImageOptionsDialog());
+        addImage2Button.setOnClickListener(v -> imageUtility.showImageOptionsDialog());
 
         // User should be able to delete a picture after they have taken it but haven't submitted the "Add item" form
         deleteImage0Button.setOnClickListener( v -> {
@@ -597,65 +588,6 @@ public class EditItemFragment extends Fragment {
 
         // Determine the appropriate ImageView to update based on the counter
         displayImages(this.imagePaths.size());
-    }
-
-    /**
-     * Displays a dialog fragment for selecting image options, such as choosing from the gallery
-     * or capturing a photo
-     */
-    private void showImageOptionsDialog() {
-        // Obtain the fragment manager for handling fragments within this fragment
-        FragmentManager fragmentManager = getChildFragmentManager();
-        // Create an instance of ImageSelectionFragment, which provides the image selection options
-        ImageSelectionFragment imageOptionsFragment = new ImageSelectionFragment();
-
-        imageOptionsFragment.setOnImageOptionClickListener(new ImageSelectionFragment.OnImageOptionClickListener() {
-            @Override
-            public void onOptionClick(int choice) {
-                // Handle the user's choice based on the selected option
-                if (choice == 1){
-                    handleGalleryIntent();
-
-                } else if (choice == 2){
-                    handleCameraIntent();
-                }
-            }
-        });
-
-        imageOptionsFragment.show(fragmentManager, "ImageOptionsFragment");
-    }
-
-    /**
-     * Handles the camera intent by checking for CAMERA permission, requesting it if necessary,
-     * and launching the camera application to capture an image.
-     */
-    private void handleCameraIntent(){
-        try {
-
-            if (ContextCompat.checkSelfPermission(requireContext(), Manifest.permission.CAMERA)
-                    != PackageManager.PERMISSION_GRANTED) {
-                ActivityCompat.requestPermissions(requireActivity(), new String[]{
-                        Manifest.permission.CAMERA
-                }, REQUEST_CAMERA_PERMISSION);
-            }else {
-                Intent cameraIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-                startActivityForResult(cameraIntent, REQUEST_CAMERA);
-            }
-
-        } catch (SecurityException e) {
-            // Handle the exception, e.g., request the permission or show a message to the user.
-            e.printStackTrace(); // Log the exception for debugging purposes.
-        }
-    }
-
-    /**
-     * Handles the gallery intent by launching the gallery to allow the user to select from the
-     * photo gallery
-     */
-    private void handleGalleryIntent() {
-        Intent galleryIntent = new Intent(Intent.ACTION_PICK, MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
-        galleryIntent.setType("image/*");
-        startActivityForResult(galleryIntent, REQUEST_GALLERY);
     }
 
     /**
